Replace autofocus logic for sending notifications
The logic to not apply autofocus for the first
textboxes in the 'send a single email or sms'
flows was stripped out in this commit:

https://github.com/alphagov/notifications-admin/commit/54b55e46c2025f747d2804dd329ed58f8dd6facd

This replaces it. It is needed because the
autofocus causes more problems than it fixes when
entering the email address or phone number but
seems ok when entering placeholders. This finding
came from a previous accessibility audit, as
described in the commit before it was removed:

https://github.com/alphagov/notifications-admin/commit/826b04282b982d36b6b0e55d467e949252b29a6d