Reorder buttons on letter template view
Fixes https://trello.com/c/7IYHVqM5/947-visual-order-of-buttons-on-letter-template-page-doesnt-match-the-html

Accessibility audit raised an issue with the focus
order of buttons on the letter template.

Cited as "The tab order is unpredictable, as it
moves in the following order:
- Get ready to send
- Add logo
- Change postage
- Edit
- Change your contact details.

It is not the expected tab order of
moving from top to bottom, left to right", it
affects keyboard-only users

This PR moves the "Get ready to send" but to
the suggested order of:

- Add logo
- Change postage
- Get ready to send
- Change your contact details
- Edit

There is extra complexity as not all buttons exist
for all users, so a decision was made to break-up
the existing `if user has "manage templates"`
block to add the button.

Tests needed to also be updated to match the
changed button order.