Improve the report title for single unsubscribe requests
It doesn’t make sense to talk about a range of time (‘until’) when
there’s only a single request in the report. As in other cases we should
talk about a specific time (‘at’).

This is something we might see as people are testing the feature and
download the report immediately after unsubscribing themselves.