Show skiplink on CSV upload page
The skiplink on the page to upload a CSV file wasn't appearing when you
tab to it, which is confusing for keyboard users who need to see the
focus indicator to know where they are in the page.

This makes it visible by giving it an opacity of `1`. The current
positioning would place the skiplink on top of the upload button on
mobile, so now only applies to bigger screens.