Remove notification id hash from the link to inbound text messages
If a service has inbound text messages, they are
able to view those messages from the the link on
the "inbox" page (services/<id>/inbox).

That link is made out of the notification id and
a hash of the latest message id, so that the link
takes you directly to that content.

This was raised as an accessibillity issue where
"This can be disorientating for blind users
because content above this
such as the h1 and other important messages may
be missed, or it may be time-consuming
for the user to orient themselves on the page.
This issue will affect screen reader users and
keyboard-only users that use the tab key to
navigate."

This PR removes the notification ID hash from the
link, so that now on page load, focus is naturally
taken to the top of the page.