Add a `Notification` model
Notifications are one of the last database objects in the admin app that
we are still passing around as raw JSON, rather than wrapping them in a
proper Python object.

This is a bit fiddly because the fields present are different depending
which API endpoint the list of notifications is coming from.

So I’ve made multiple models instead of trying to overload one.