Fix split get_notifications_for_service and add test
The split had to be reverted because it introduced a bug
- for jobs, there is a different function in the API repo
than for all service notifications, and that needs
format_for_csv flag to be passed in, but we have not done
that., hence the bug.

This commit reintroduces the changes, with a fix.

It also adds a test that guards that the same issue couldn't
happen again.