Let user choose which version of Notify docs they want to view
When sending them a link to a particular section in the docs.

We do this as often when sending service users links to a section
in the docs, we don't know which of our clients they use, if any.

I also included "Any is fine" option to choose, for users who
aren't very technical, but would still like to see the docs.
It might be hard for these users to choose from a list of options
they are not familiar with. Hence, I provided a "comfort" option
for them. Selecting this will take them to rest API docs.