Change folder paged heading issue description
We're now pretty confident that JAWS was splitting
the headings up because they contained more than a
set number of characters.

You can change how many characters JAWS will break
after so this can be configured. We are now
instead focusing our work on changing the design
of the page heading so better to rewrite this to
talk about the general effect on users rather than
being exact about the experience with JAWS set up
in a certain way.