Introduce a Rollup config
This PR introduces a Rollup config to manage
various build steps

https://rollupjs.org/configuration-options/

Config contains three separate sections:

1. ESM JavaScript bundle compilation
2. SCSS compilation
3. Legacy Javascript compilation

1. ESM JavaScript bundle compilation

The simplest one. Single input to create
an ESM bundle that contains sourcemaps only in
development and a minified file for production.

As Rollup "task" requires and input anc output,
and the copy plugin doesn't have separate ones,
it's been added to this one.

2. SCSS compilation

Sadly, globbing isn't supported so we need to
list the entry points we import.

We minimise for production, include sourcemaps for
development and we disable the CSS URL resolver
warnings (as referenced static files exists
elsewhere). For production build we also rewrite
the url to those assets as they're stored elsewhere.

3. Legacy Javascript compilation

We use multi entry to first import
external dependecies, then local files and finally
`main.js` - same as before.
For development we include sourcemaps and for
production we minify.

To prevent `this is undefined` error we apply
`window` context for `jquery` and `cbor` via
[`moduleContext` Rollup option](https://rollupjs.org/troubleshooting/#error-this-is-undefined)

Additionally we needed to not mangle "Hogan" in
minification as the script sets a global variable on
`window` that is then used in one of our scripts.