Show additional button text on mobile
Currently we visually hiden some button text as
buttons are near the area of edit.
With teh change on mboile to move all the buttons
above the preview, we need to show that text as
it doesn't make sense without it.

To do that, we replace the `govuk-visually-hidden`
CSS class with our custom one and use the
`govuk-visually-hidden` mixin to visually
hide the text on larger viewports.