EventletTimeoutMiddleware: wrap application.wsgi_app, get value from env var
wrapping `application.wsgi_app` is more correct than wrapping
`application` itself, as it preserves flasky behaviours of the
outer `application` object

also control the timeout time with a
`HTTP_SERVE_TIMEOUT_SECONDS` config/env var that we can control
from e.g. terraform, unifying this with the value we use for
gunicorn's `timeout` (which has little effect on the eventlet
`worker_class`, but controls a similar behaviour on non-async
workers), with the intention we start to respect this across all
our apps.

also default to 30s is what we've decided to set our cloudfront
timeouts to.