Change H1 and title on page where email reply-to address is checked
When adding an email reply-to address, the page to check the addres had
an H1 of "Add email reply-to address" which is the same as the H1 on the
page before (the page to enter the address). Having the same heading and
title on consecutive pages makes it unclear to screen reader users that
the page has changed, so changes the H1 and title for the page where we
check the address.