Update service navigation to show names
CSS changes to allow

- display full name of the organisation
(if allowed to see)
- display full name of the service regardless of
the viewport size or text zoom level.
For this we mimicked the gov.uk breadcrumb pattern,
where a long item drops onto the next line with a
chevron preceding it.

Why
Organisation and service names were being
truncated if they were too long of their container,
which fails 1.4.10 Reflow.

Changes in this PR:

- Remove overflow hidden and no-wrap
We need to display the full contents of the org
and service name
- Remove max widths
We want the containers to take up as much space
as the contents within them requires
- Use flexbox to position items
Settings flexbox on `navigation-service` container
allows us to set a desired width on the new
wrapping container for org and service name as
well as remove absolute positioning on the
`switch-service` link to fix content overlap.
- Move service nav chevron to service name container
We want the chevron to move into the next line
with the service name when organisation and/or
service names are long. Apply chevron only when
both org and service name present as we do not
want the chevron on the service name if there is
no org link, so we make use of next-sibling
combinator to shown chevron and add left-padding.
- Add bottom margin to organisation link
this is to give a bit more space to the service
name when it wraps (especially for the focus state)
- Float items
Floating instead of inline-block removes the
gap between items without resorting to the
zero font-size hack.