Position and focus tweaks
1. remove block focus state for organisation and switch service links

As content now wraps we can no longer have a block focus state on the organisation link. And as the org link is no longer block it looks weird for the "switch se"rvice link have have a block focus state.
"Back to" link that is there on content pages (when signed in) still has a block focus state.

2. Add bottom margin to organisation link

this is to give a bit more space to the service name when it wraps (especially for the focus state)

3. Adjust chevron position based on new padding values

As content can be multiline and font size changes on tablet and up, we adjust it a bit.