Fix sticky footer overlapping some focused elements
Adjust "scrollToRevealElement" function, adding
a condition that checks for the type of
focused element and if it's either a checkbox or
radio then take the parent element (which will be
`govuk-checkboxes__item` or `govuk-radios__item`)
and adjust the scroll for its height.

This fixes an issue where checkboxes on templates
page were not brought into view when focused if
the sticky footer happens to overlap them.

While that's due to them having a minimal height
for visual composition, it also helps to now
fully show the hint text as that is usually lower
than the height of the checkbox itself.