Remove classesPersister JS and associated code
The classesPersister JS was a slight hack we
introduced in the following pull request:

https://github.com/alphagov/notifications-admin/pull/4155

It was added to make our 'block focus style' work
with the variant of template list items emergency
alerts introduced. In this variant, the link is
wrapped in a <h2> tag which stopped our existing
CSS targeting sibling elements.

Emergency alerts code was removed in this pull
request:

https://github.com/alphagov/notifications-admin/pull/4922

Because there is now no HTML that uses this code
in our app, we shouldn't include it and be sending
it to users. This removes it.