Fix sticky footer overlapping some focused elements
Adjust "scrollToRevealElement" function, adding
a condition that checks for the type of
focused element and if it's either a checkbox or
radio then take the parent element (which will be
`govuk-checkboxes__item` or `govuk-radios__item`)
and adjust the scroll for its height.

This fixes an issue where checkboxes on templates
page were not brought into view when focused if
the sticky footer happens to overlap them.

While that's due to them having a minimal height
for visual composition, it also helps to now
fully show the hint text as that is usually lower
than the height of the checkbox itself.

Update stick-to-window-when-scrolling tests
- Wrap checkboxes in containing DS elements,
like they appear on the website as the script
adjusts the offset by their height

- Mock position of the checkbox parent on
"stick at top" scenario to allow
the test to scroll the window enough

Check for isInSticky on focus
If radio or checkbox is in sticky
there is no need to run the code.

We can't move the `if (isInSticky()) { return; }`
to the top, as then the code  in `else` block
runs which we don't want.