make view_notifications ajax blocks have same URL/method
The URLs were different for the two ajax requests due to a missing
pagination argument - as a result, we ended up with two different
entries in the list of ajax calls, with two different requests to update
their respective blocks. The ajax endpoint returns both blocks and
updateContent.js is designed so that if two blocks request ajax content
from the same URL, it'll just make one remote call and call the renderer
for each block when it gets a response.

So to fix this, we could just add the pagination argument. But we also
need the form to be defined in this block too - otherwise we'll end up
in a weird and confusing state where one ajax block will make a GET
request without the form search data, and one will make a POST request
with the form search data (thus doing additional filtering). The first
one to return would redraw both methods as it empties the queue. But
both would also put on the timeout.

So if you have a GET and a POST with the same URL you could have
theoretically end up with states where you were flip-flopping between
filtered and non-filtered results - that said I think this wouldn't
happen in practise due to the caching of the search results that we
added recently.