Put email metadata in summary list, not a table
It's 1-dimensional data, just representing a
mapping of a field's key ('From') to its value
('Tom's service') so should be in a summary list,
which the design system guidance, and the HTML5
spec says can be used for metadata:
- https://design-system.service.gov.uk/components/summary-list/
- https://html.spec.whatwg.org/#the-dl-element

We often get it flagged as a problem in
accessibility audits that we're showing
1-dimensional data in tables. The report from the
most recent one described an auditor being
confused by the inclusion of a table here. That
was mainly because it didn't have a caption but
this will also be because tables imply a certain
type of data, which this isn't.

Notes on changes:
- on smaller screens, the current layout has too
  much side padding so this reduces that
- the most ideal way to layout this content is
  with CSS Grid (using subgrid) so we use it here,
  targeted to browsers that support it only and a
  reasonable fallback for those that don't