Don’t coerce `row_number` in CSV download
The `row_number` that the API returns for a CSV download can be either
an empty string or a number.

When we deserialise the JSON from the API we check for `None` values
before trying to cast to another type, but we don’t check for empty
string.

Therefore marking a field as `int` causes an exception because empty
string cannot be coerced to an integer.

This commit change the type to `Any`, which means no coercion will be
attempted.