Give checkboxes height again & fix other bugs
Includes changes to the template list CSS to:
- give the checkbox element has height again
- fix an issue caused by design system changing
  the `box-decoration-break` value to assume all
  lines a link wraps to inherit the initial
  padding
- add padding-left to stop content from the
  template list item wrapping underneath the
  checkbox

The checkboxes currently have a height of 1px,
which interferes with the logic browsers use to
scroll them into view when mostly off-screen.
Because only their first 1px was considered by the
layout, you would often see checkboxes that were
mostly off-screen focused without being scrolled
into view.

The `box-decoration-break` change interferred with
us having padding at the start of links to
folders, to make space for the folder icon. It was
only visible when the links were focused.