Add back link to "Choose a template" page for inbound SMS
When choosing to reply to an inbound SMS you get taken to the "Choose a
template" page which had no back link. This adds one.

The back link will go to the parent folder (if looking at templates in a
sub-folder) or to the inbound SMS conversation page if already looking
at the top-level templates.