Make back links go back to inbound SMS flow if sending from this route
If sending a text message from the inbound SMS flow (by viewing an
inbound SMS and then choosing to reply to it) the back links should take
you back to that flow, and not back to "standard flow" of coming from
the templates page.

This change adds a key to the session, `from_inbound_sms_details`, if
coming from the inbound SMS flow. The presence of this key and the
values it contains shows which page to go back to if you're sending an
SMS and are on step-1 of the send flow. We clear this key from the
session once a notification is sent so that it doesn't affect future
journeys.