Give all folder icons alt text
Their purpose in the visual interface is to show
the thing they are next to is a folder. We need
this information in the semantic interface too, so
users of assistive tech can access it.

Note: this doesn't change the folder on the team
member page because it adds no extra information
to the following text ('X of X folders').