Expand table macro to accept header width classes
On the inbound sms inbox page we want the table
columns to be 2/3 and 1/3 wide.
To do that, a width can be set on the table header
cells.
We can do that by extending the macro to either
accept and arry of string of table headings
(current) or an array of objects containing both
the header text with `text` property and HTML
classes with the `classes proprty`.
This could have been done with supplying the
`width` values instead, but this approach should
make it easier to migrate to the Design System
table component eventually.

For width we created 2/3 and 1/3 variants of the
Design System `govuk-\!-width-<width>` suffixed
with `--static`, as unlike the Design System ones
that change to 100% on smaller vieports, we want
them to remain as is.

This extended macro is now used on the inbound sms
inbox page and it includes the test to check their
classes are present.