Conditionally focus template page forms
When a user clicks say "add a new template" button, we show and focus that form.
Should then a user submit that form without any selection, we reload the page, show the error-summary banner, check which form they submitted and then show it and focus it.

This causes issues for screen render users, which expected the error summary to be focused.

This PR amends the javascript module logic to check if the rendered page contains an error summary. If it does, we don't focus the form and let Design System Javascript handle error-summary.

As the error-summary contains the link to the errored form, users can then make changes quickly.