Apply extra HTML class to error-summary on /templates page
On mobile view the error-summary is right against
the service navigation, as there is no back-link
like on other journeys.

To set a margin top, we extended the `errorSummary`
macro in the `withnav_template` to accept
`error_summary_extra_params` parameter from the
`templates.py  > choose_template` function,
passed via the `render_template`

`error_summary_extra_params` is an object so any
parameters that the error summary nunjucks macro
accepts can be passed.

In the `errorSummary` macro we already set some
default parameters. If any extra are passed,
we join them together and pass it to component
to render.