Send auth type from service join request choose permissions form to API
The SMS or email auth radio buttons were already being shown on the form
to choose a new users permissions if a service had email auth enabled,
but weren't doing anything with the data selected.

This makes the form work in the same way as the form to change a user's
permissions. We only show the radio buttons if a service has email auth
enabled. And if the service has email auth enabled but the user has not
got a mobile number then we make the sms option not selectable.