Update templates template to conditionally show the notification banner on successful PDF attachment upload
Currently, when an PDF attachment is successfully
uploaded to a template, we redirect to a template
preview and anchor to the first page of the
attachment.

Visually that shows that the attachment was
successfully uploaded (we do show an error if it
wasn't), but for screenreader users do not get
informed.

This PR attaches a `upload_successful="yes"` query
string to the URL, based on which we
render the Notification banner component -
Success variant -
https://design-system.service.gov.uk/components/notification-banner/

The template will then show the notification
banner before the H1 with an anchor link to the
first page of the attachment that already exists.

This way we indicate that the upload was
successful for both visual and assisstive tech
users.