Split out breadcrumb part of headings with paths
We've heard that having the full breadcrumb in the
heading is confusing for screen reader users,
mainly because it's odd to have several separate
elements in your heading but also because
breadcrumbs are a known UI pattern and don't
generally exist in headings.

This splits the path to the folder out from the
heading. Visually, it will look the same which
doesn't totally solve the problem with the
expectations around breadcrumbs but the heading is
at least visually different from the rest of the
path, most of the time.