Update estimate-usage page
Use checkboxes with conditional content to display input fields.

Field validation remains as it was. If fields have value or if they contain the error we check the checkbox so that the field is display and error visible (if applicable)

Prototype code as checkboxes are in the view and we don't validate if the submit without checking any.