Show email sender name page as part of sending an email
The page to select an email sender name now gets shown after a user adds
their first email template if they haven't already confirmed which email
sender name they want to use.

The back links on this page go to either the settings page or the
templates page, depending on how you got to the page.