Limit titles for template/folder pages to 4 parts
We currently list out the full path. This instead
just includes the current template/folder and its
parent, along with the service and 'GOV.UK
Notify'. The idea is to ape the format used by
many websites, including www.gov.uk, where the
assumption is that including the parent gives
enough context to the page when the name of the
page might be shared by others in the same domain,
as with us.

This breaks with that format by also including the
service name, because that seems to be vital
context, specific to Notify.