Fluid big number sizes with clamp()
Use clamp to create a somewhat fluid big number
scaling between a minimum and maximum

To prevent long number breaking out of the
container we use a combination of clamp -
https://developer.mozilla.org/en-US/docs/Web/CSS/clamp
and container query values -
https://developer.mozilla.org/en-US/docs/Web/CSS/CSS_containment/Container_queries
to scale the font-size as the container size changes.

Mixin accept a size from the DS type scale -
https://design-system.service.gov.uk/styles/type-scale/#the-old-type-scale
- and uses their function `govuk-px-to-rem`
to generate REM values to be used inside `clamp()`

Additionally, the mixin allows for bold and tabular
number (if it gets used elsewhere)

Container query properties and values used on
a visual basis to achieve teh desired affect.

On smaller screens and for
IE11 we use the minimum value.