Change links to and from email reply-to route from go-live page
If you get to the page to add a reply-to address from the go-live
checklist, the back link on this page takes you back to the go-live
checklist (instead of the default of the settings page).

After you have added a reply-to address from the go-live checklist we
now redirect you back to the go-live checklist (instead of the page to
show all email reply-to addreses).

If you are on the page where the email address is still being verified
and click "back" twice you will be taken out of the go-live route, but
this is an edge case and fine for the prototype.