Change links to and from email reply-to route from sending journey
If you get to the page to add a reply-to address from the route of
sending an email, the back link takes you back to the page to set the
sender.

After you have added a reply-to address as part of the route of sending
an email, you get taken to the page to send the email (if you've just
added the first address) or to the page to set the sender (if you have
multiple senders).