Refactor displayed postage
By moving this into the model we move a bunch of branching logic out of
the view layer and make it easier to potentially reuse.

Displayed postage is what we use to decide what to show in the corner of
the envelope.

If the letter can’t be posted (because validation failed) we don’t want
to show any postage, becuase the letter will not be sent.

Other parts of the code still need access to `.postage` to calculate
delivery times etc.