Remove fixed issues from acessibility statement
One of these was a left over from the last time
this statement was updated so was actually fixed a
while ago.

Also includes a change to our listing of the issue
with some h1s. The fix we made is due to be tested
in the next few months so while the redesign will
definitely help with some aspects of the issue
reported, we can't confirm it solves the confusion
reported until we test again with accessibility
auditors.