Make success banners alert screen reader users
Tested with NVDA and JAWS and the banner gets
announced when the page loads with the page title
following.

We had concerns that giving success banners, which
usually appear after a redirection, the role of
alert would stop the page title being announced.
This was unfounded and they do not so this change
is a sensible way of making sure screen reader
users are notified when an action is successful.