Always show headings on the "Your Services" page
There are four potential sections on the "Your services" page:
- Platform admin
- Organisations
- Live services
- Trial services

We were showing the section headings if you were a platform admin user,
or if there were more than two visible sections. We now want the live
and trial sections to always show (with set text if they are empty).
This means we can get rid of the logic which decides whether or not to
show the section headings since we will always be showing at least two.