jobs: make aware of "finished all notifications created" job status
while still working correctly without it so that we can roll this
out before api support - only minor test changes in this commit
in an attempt to prove that the only observable difference with
this code change is a new status included in filter when trying
to list jobs. this should be ignored by the api endpoint.