Give content pages' sub nav padding-bottom
This makes sure it has the same distance from the
page footer as the main content column.

Includes a whitespace fix from ruff and the
addition of '!important' to the existing styles
for this padding style. '!important' makes the
browser ignore the order of precedence and just
apply this style over any others for
padding-bottom on this element. We would normally
never use '!important', following best practices,
but do so here to follow design system's design
for the override classes we're extending.

See:
- https://developer.mozilla.org/en-US/docs/Web/CSS/important#best_practices
- https://design-system.service.gov.uk/styles/spacing/#overriding-spacing