Rewrite homepage JS as an ES Module
We want to migrate all of our legacy Javascript
files to ESM, so this is the latest in the series of conversions.

Rewrite follows the same conventions we
previously used for FocusBanner, Collapsible
checkboxes, ColourPreview, FileUpload and
Autofocus.

Added bonus: removes the use of jQuery as well

In this PR we:

- rewrite the module as ES Module
- remove old module JS file
- remove old module from the legacy bundle build