Sync newly-frozen requirements
If we don’t do this then we’ll get old versions of the config files
from utils, and we won’t get newer versions of any updates
subdependencies installed.

Safe to assume that if someone is freezing requirements it’s because
they want to use those new requirements.

Matches what we’ve done in the API repo here:
https://github.com/alphagov/notifications-api/commit/655bccda9688ebd30dca8c8c7e11494215cc5169