Stop passing the year to API when setting SMS fragment limit
The year argument was not being used by the api, and we don't let you
specify which year you are setting the free SMS fragment limit for
anyway - you can only set it for the current financial year.

This also updates a mock to make its return value match the real format.