Remove .ac.uk from the generic list of domains
We are not letting any new services go live from universities. To make
this less likely, we should stop anyone with a university email address
from creating an account automically. If there are exceptions they can
still contact us and ask for their specific organisation and domain to
be added to our database.