Don’t let platform admins try to make archived services live
A platform admin user can try and make an archived service live - we
would not do this intentionally, but it could happen if someone clicks
the button to make the service live just after the service was archived.
If this happens the user will see the “something went wrong” error page
due to API giving a validation error of
```json
{'normalised_service_name': ['Unacceptable characters: `normalised_service_name` may only contain letters, numbers and full stops.']}
```

It shouldn’t be giving this error, it should be clear what has happened
instead.

I think a ‘permission denied’ error is clear enough for this edge case
– if the user goes back and tries again they will see that the service
is now archived.