Rewrite ListEntry as ES Module
As we want to migrate all of our legacy Javascript
files to ESM, this is a small part.

Rewrite follows the same convention we previously
used for FocusBanner, Collapsible checkboxes,
ColourPreview and others:

https://github.com/alphagov/notifications-admin/pulls?q=is%3Apr+is%3Amerged+%27as+an+ES+Module%27

This also includes the removal of all uses of
jQuery, as a bonus. Otherwise, the code should be
pretty much unchanged, just moved to the new
syntax.