Give banner role of 'status' if just an update
We currently use the role of 'alert' for all
banners. In recent accessibility testing, the
Digital Accessibility Centre (DAC) suggested we
should only be using that role for important or
time-sensitive updates. This changes the banner
component so the default role is 'status' to
reflect that.

In practice, this will change how the green
banners are announced by screen readers.