Move the status update into the main content
We currently use the role of 'alert' for all
banners. In recent accessibility testing, the
Digital Accessibility Centre (DAC) suggested we
should only be using that role for important or
time-sensitive updates. They instead suggested
using a role of 'status' and not to focus it. This
would result in the update being announced only
when other announcements are complete.

We also need the page to reflect whatever the
current state of the address check is. It kind of
does now but there's no content between the h1 and
'Continue' button, which is confusing, because
it's in the success banner.

This is important because screen readers can
easily miss status updates so will instead rely on
reading the page content to get what's happening.

Talking to @karlchillmaid, we decided to move the
status update inside the main content area and
rewrite it to better follow what was there before.
I'm also wrapping it in a live region, like the
success banner was, so that the changes will get
announced.

It's worth saying that users will be taken back to
the Reply-to email addresses page when they click
'Continue' so will be able to see the address they
added in the list as further confirmation, if
successful.

This has been tested in NVDA, JAWS and Voiceover
and works as described here.