gunicorn: reduce keepalive timeout to 35s
this was originally set to a generous value >60s, because we
had our ALB idle timeout set to 60s and ALBs are known to
have problems if the keepalive connection isn't held open
for more than the idle timeout. now that we've reduced the
ALB idle timeout to 30s, we should be able to reduce the
keepalive timeout too.

excessively long keepalive timeouts cause a GreenThread (and
all its associated thread-local resources) to be unnecessarily
reserved for that period of time following a request, no
matter how short the original request.