Fix overflowing email addresses in template preview
When previewing the email template befoire sending, any long email adresses in "Reply to" or "To" fields will overflow the container on mobile or high zoom level.

We already had a fix for the "reply-to" fields.
This adds the same fix to the "To" filed by adding a HTML class to the list item and includes it in the CSS style rule.