Add error prefix to attach pages page
We normally rely on the admin_template.html jinja
to add an 'Error:' prefix when validation errors
occur. This page does its own validation, sending
the errors through in an 'error' dict instead.

We added the error summary when we made this page
but didn't add the prefix. This does that.

This behaviour follows the 'Recover from
validation errors' pattern from the design system:

https://design-system.service.gov.uk/patterns/validation/#how-to-tell-the-user-about-validation-errors