Make edit guidance say ‘send files’ not ‘send documents’
When editing a template the guidance talks about ‘sending documents’.

This language is outdatd, because there are now various types of file
you can send using Notify.

On the man guidance page[1] we talk about sending ‘files’ rather than
documents. This commit updates the in-page guidance on the edit template
page accordingly.

1. https://github.com/alphagov/notifications-admin/blob/4ab63f95be583f1deafee8772a57ef97cac357e6/app/templates/views/guidance/using-notify/send-files-by-email.html