Add endpoint to download report requests
This endpoint gets the report details from the api so that we can check
the user has permission to download it and it is in the right state. The
api will only return reports which match both the service ID and report
request ID, so will not return a report request for another service the
user is a member of.

If the report is fine to download, we then get the file from S3.