Add endpoint to download report requests
This endpoint gets the report details from the api so that we can check
the user has permission to download it and it is in the right state. The
api will only return reports for the user returned by the api, and is
currently a platform admin only view. When we want to roll this out, we
can change that so the report can be downloaded by users with the
`view_activity` permission.

If the report is fine to download, we then get the file from S3.