Provide message status in the visually hidden heading
We want to convey to SR users what the status of
the message that follows the heading is. To do that,
we append the a friendly status text.
We also make the visible status text and date
container not keyboard accessible so as to not
duplicate annouced content.

Reworked message status conditional to set
heading text as well as visible sttaus text in one
go.