Let platform admins remove each other’s permission
Sometimes we need to do this if someone has lost their Yubikey and needs
to log back in to add a new one.

It might also be useful if we ever needed to revoke someone’s platform
admin access in a hurry (although we’d probably archive their whole
account in that case).

Keeping some safeguards in here so that it shouldn’t be possible to
write code which grants platform admin access through the UI (that
should still be done through the DB).