Add form for users to request CSV reports
For generating async reports we need to first
create a request entry. We do this by wrapping
this in a form that POSTs the data.

As we want the button to look like then link and
it to be aligned with the retention text, the
markup needs to be updated as our render doesn't
like a 'form' to be in a paragraph.

New form is for now behind a platform_admin flag.