Create 'preparing your report' page
Holding page while backend is generatign a report
Main content is inside a container with a
role=status as we augment this page with Javascript
to check if the report status has been updated.
If it has we update the page text so that it is
announced to screenreaders.

For no-JS there is a a link that reloads the page
and application logic then handles it.