CSS styles for button as link and inline form
We don't want the request CSV report form button
on the notifications page to look like the button
due to it's proximity to the search box.
We're overriding button's hover, focus and active
states and applying link text styles.

To have the form and retention text side by side
we're extending the DS 'govuk-button-group' CSS
class and introducing 'form--inline' CSS class
to maske it all work with flexbox.